solution/backend/microservices/cybernetics_core/requirements.txt
flask==3.1.0
flask_cors==5.0.0
pydub==0.25.1
groq==0.15.0
azure_cognitiveservices_speech==1.42.0
dependency-injector==4.45.0
langchain-groq==0.2.3
pytz==2024.2
pycountry==24.6.1